+++ src/postulo/core/server_urls.py
@@ -17,6 +17,7 @@ urlpatterns = [
     path("email/test/", views.EmailTestView.as_view(), name="email_test"),
     path("plugins/", views.PluginsView.as_view(), name="plugins"),
     path("plugins/action/", views.PluginActionView.as_view(), name="plugin_action"),
+    path("logs/", views.LogsView.as_view(), name="logs"),
     path("capture/", views.CaptureView.as_view(), name="capture"),
     path("defaults/", views.DefaultsView.as_view(), name="defaults"),
 ]